assignment 5 part 2
