+++ problems/fizzbuzz/coopersenior/index.html
@@ -46,7 +46,7 @@
     </p>
     <h2>Navigation</h2>
     <ul>
-      <li><a href="../../index.html">Back</a></li>
+      <li><a href="../index.html">Back</a></li>
     </ul>
     </p>
 </body>
