+++ problems/fizzbuzz/index.html
@@ -41,7 +41,7 @@ FizzBuzz
     <ul>
       <li><a href="C#/CarleeJoppa/index.html">C#/CarleeJoppa</a></li>
       <li><a href="python/ryangross/index.html">python/Ryan Gross</a></li>
-      <li><a href="python/coopersenior/index.html">python/Cooper Senior</a></li>
+      <li><a href="coopersenior/index.html">Cooper Senior</a></li>
     </ul>
     </p>
     <p>
